examples/python_find/requirements.txt
paho_mqtt==2.1.0
pymongo==4.10.1